fix: size_t
[skip changelog]

+++ src/MaaCore/Config/Miscellaneous/OcrPack.cpp
@@ -133,7 +133,7 @@ asst::OcrPack::ResultsVec
     }
     else {
         std::string output = "[";
-        for (int i = 0; i < raw_results.size(); ++i) {
+        for (size_t i = 0; i < raw_results.size(); ++i) {
             if (i != 0) {
                 output += ", ";
             }
